requirements.txt
mysql-connector-python
fpdf2